% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotus_filtered}
\alias{scotus_filtered}
\title{Random sample of Supreme Court of the United States opinions with more than 1000 characters}
\format{
A data frame with 10000 rows and 5 variables.
}
\usage{
scotus_filtered
}
\description{
Random sample of Supreme Court of the United States opinions with more than 1000 characters
}
\keyword{datasets}
