% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotus_sample}
\alias{scotus_sample}
\title{Random selection of Supreme Court of the United States opinions}
\format{A data frame with 6336 rows and 5 variables.}
\usage{
scotus_sample
}
\description{
Random selection of Supreme Court of the United States opinions
}
\keyword{datasets}
